+++ src/FluxFlow.Components.Assertions/Options/AssertionsComponentOptions.cs
@@ -71,8 +71,12 @@ public sealed class AssertionsComponentOptions
         ArgumentException.ThrowIfNullOrWhiteSpace(name);
         ArgumentNullException.ThrowIfNull(type);
 
-        _types[name.Trim()] = type;
-        _types[type.FullName ?? type.Name] = type;
+        lock (_typesLock)
+        {
+            _types[name.Trim()] = type;
+            _types[type.FullName ?? type.Name] = type;
+        }
+
         return this;
     }